requirements.txt
django
djangorestframework
celery
redis
django-celery-results
django-cron
django-filter
environs